Remove deprecated `dymanicPatterns` in known issues
