+++ vividus-soft-assert/src/main/java/org/vividus/softassert/issue/KnownIssueIdentifier.java
@@ -16,7 +16,6 @@
 
 package org.vividus.softassert.issue;
 
-import java.util.Collections;
 import java.util.Map;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -31,11 +30,7 @@ public class KnownIssueIdentifier
     private boolean failTestCaseFast;
     private boolean failTestSuiteFast;
     private String description;
-    private Map<String, Pattern> additionalCompiledPatterns = Collections.emptyMap();
-
-    @Deprecated(since = "0.3.10", forRemoval = true)
-    private Map<String, Pattern> dynamicCompiledPatterns = Collections.emptyMap();
-
+    private Map<String, Pattern> additionalCompiledPatterns = Map.of();
     private Map<String, Pattern> runtimeDataPatterns = Map.of();
 
     public KnownIssueType getType()
@@ -109,23 +104,10 @@ public class KnownIssueIdentifier
                 .collect(Collectors.toMap(Map.Entry::getKey, e -> Pattern.compile(e.getValue(), Pattern.DOTALL)));
     }
 
-    @Deprecated(since = "0.3.10", forRemoval = true)
-    public void setDynamicPatterns(Map<String, String> dynamicPatterns)
-    {
-        this.dynamicCompiledPatterns = dynamicPatterns.entrySet().stream()
-                .collect(Collectors.toMap(Map.Entry::getKey, e -> Pattern.compile(e.getValue(), Pattern.DOTALL)));
-    }
-
     public void setRuntimeDataPatterns(Map<String, String> runtimeDataPatterns)
     {
         this.runtimeDataPatterns = runtimeDataPatterns.entrySet().stream()
                 .collect(Collectors.toMap(Map.Entry::getKey, e -> Pattern.compile(e.getValue(), Pattern.DOTALL)));
-    }
-
-    @Deprecated(since = "0.3.10", forRemoval = true)
-    public Map<String, Pattern> getDynamicCompiledPatterns()
-    {
-        return dynamicCompiledPatterns;
     }
 
     public Map<String, Pattern> getRuntimeDataPatterns()
